Round a float64 tensor in binary64 in the elementwise quantizers (phase 3)
binaryK_quantize and superfp_quantize narrowed a float64 operand to float32,
rounded it there and widened the result back: two extra passes, two
temporaries, and an input float32 could not hold was rounded twice (RNE of
1.0625 + 2^-30 to P = 4 gave 1.0 where the input is above the tie and owes
1.125). Both now round it once, in binary64, on both devices.

The functors, params and casts take carrier_t<scalar_t> (bit_helper.h: double
for float64, float for every other dtype), the elementwise dispatch gains
float64, and the narrow/widen around the four entry points is gone; the GEMM
keeps narrowing until phase 4. A float64 element under SR reads two of a
Philox block's four words, so one block covers one double2 load
(philox.h's word64; element j uses block j >> 1). The float word layout,
and every float loop, is untouched: the double SR loops are separate
if-constexpr arms in cuda/utils.cuh and cpu/utils.h.

MPTORCH_NO_FP64 is not applied here: the four TUs grow by 0.4-2.5 s
(wall 50.0 -> 50.9 s, .so +326 KB), not worth separate _f64 files.

128 of 128 existing device kernels are SASS-identical; the 14 double
kernels use 30-46 registers, no stack spill. 1126 of 1152 host bodies
unchanged; the 26 that moved are the entry points, their .cold clones,
the CUDA registration and host launchers. Float captures bit-identical;
the float64 SR rows moved, as the plan predicted. short, image64-short and
sr-mean pass. tests/test_quantize_dispatch.py is rewritten around the new
contract -- float32-valued float64 gives the float32 image in every
deterministic mode, inputs float32 cannot hold round once, formats only
binary64 can carry (P up to 53, bias up to 1022, via the raw op until
phase 5 lifts the checks) match the P3109 transcription, SR draws are
keyed on the index, neighbours only, unbiased -- and 54 of its 128 tests
fail on the previous build. pytest 2856 passed.

Float rows new/old: CUDA 0.999, CPU 1.007 (t1) / 1.005 (t8). float64 rows:
CUDA 0.51, CPU 0.62 (t1) / 0.60 (t8).

+++ mptorch/csrc/cuda/simd_traits.cuh
@@ -19,6 +19,8 @@ struct SIMDTraits<float>
     }
 };
 
+// The one width whose lanes are not handed to `f` as binary32 values: a double
+// rounds in binary64 (carrier_t, bit_helper.h), so `f` takes the double itself.
 template <>
 struct SIMDTraits<double>
 {
@@ -28,8 +30,8 @@ struct SIMDTraits<double>
     {
         const double2 *in = reinterpret_cast<const double2 *>(&v);
         double2 out;
-        out.x = f(static_cast<float>(in->x), 0);
-        out.y = f(static_cast<float>(in->y), 1);
+        out.x = f(in->x, 0);
+        out.y = f(in->y, 1);
         return *reinterpret_cast<float4 *>(&out);
     }
 };
